+++ lib/model/settings/enums/coordinate_format.dart
@@ -8,15 +8,17 @@ import 'package:latlong2/latlong.dart';
 extension ExtraCoordinateFormat on CoordinateFormat {
   static const _separator = ', ';
 
-  String format(BuildContext context, LatLng latLng, {bool minuteSecondPadding = false, int dmsSecondDecimals = 2}) {
+  String format(BuildContext context, LatLng latLng, {bool minuteSecondPadding = false, int? dmsSecondDecimals}) {
     final text = formatWithoutDirectionality(context.l10n, latLng, minuteSecondPadding: minuteSecondPadding, dmsSecondDecimals: dmsSecondDecimals);
     return context.applyDirectionality(text);
   }
 
-  String formatWithoutDirectionality(AppLocalizations l10n, LatLng latLng, {bool minuteSecondPadding = false, int dmsSecondDecimals = 2}) {
+  String formatWithoutDirectionality(AppLocalizations l10n, LatLng latLng, {bool minuteSecondPadding = false, int? dmsSecondDecimals}) {
     switch (this) {
       case CoordinateFormat.dms:
-        return toDMS(l10n, latLng, minuteSecondPadding: minuteSecondPadding, secondDecimals: dmsSecondDecimals).join(_separator);
+        return toDMS(l10n, latLng, minuteSecondPadding: minuteSecondPadding, secondDecimals: dmsSecondDecimals ?? 2).join(_separator);
+      case CoordinateFormat.ddm:
+        return toDDM(l10n, latLng, minutePadding: minuteSecondPadding, minuteDecimals: dmsSecondDecimals ?? 4).join(_separator);
       case CoordinateFormat.decimal:
         return _toDecimal(l10n, latLng).join(_separator);
     }
@@ -29,6 +31,19 @@ extension ExtraCoordinateFormat on CoordinateFormat {
     final lng = latLng.longitude;
     final latSexa = _decimal2sexagesimal(lat, minuteSecondPadding, secondDecimals, locale);
     final lngSexa = _decimal2sexagesimal(lng, minuteSecondPadding, secondDecimals, locale);
+    return [
+      l10n.coordinateDms(latSexa, lat < 0 ? l10n.coordinateDmsSouth : l10n.coordinateDmsNorth),
+      l10n.coordinateDms(lngSexa, lng < 0 ? l10n.coordinateDmsWest : l10n.coordinateDmsEast),
+    ];
+  }
+
+  // returns coordinates formatted as DDM, e.g. ['41° 24.2028′ N', '2° 10.4418′ E']
+  static List<String> toDDM(AppLocalizations l10n, LatLng latLng, {bool minutePadding = false, int minuteDecimals = 4}) {
+    final locale = l10n.localeName;
+    final lat = latLng.latitude;
+    final lng = latLng.longitude;
+    final latSexa = _decimal2ddm(lat, minutePadding, minuteDecimals, locale);
+    final lngSexa = _decimal2ddm(lng, minutePadding, minuteDecimals, locale);
     return [
       l10n.coordinateDms(latSexa, lat < 0 ? l10n.coordinateDmsSouth : l10n.coordinateDmsNorth),
       l10n.coordinateDms(lngSexa, lng < 0 ? l10n.coordinateDmsWest : l10n.coordinateDmsEast),
@@ -52,6 +67,22 @@ extension ExtraCoordinateFormat on CoordinateFormat {
     final secText = NumberFormat('${'0' * (minuteSecondPadding ? 2 : 1)}${secondDecimals > 0 ? '.${'0' * secondDecimals}' : ''}', locale).format(sec);
 
     return '$degText° $minText′ $secText″';
+  }
+
+  static String _decimal2ddm(
+    double degDecimal,
+    bool minutePadding,
+    int minuteDecimals,
+    String locale,
+  ) {
+    final degAbs = degDecimal.abs();
+    final deg = degAbs.toInt();
+    final min = (degAbs - deg) * 60;
+
+    final degText = NumberFormat('0', locale).format(deg);
+    final minText = NumberFormat('${'0' * (minutePadding ? 2 : 1)}${minuteDecimals > 0 ? '.${'0' * minuteDecimals}' : ''}', locale).format(min);
+
+    return '$degText° $minText′';
   }
 
   static List<String> _toDecimal(AppLocalizations l10n, LatLng latLng) {