+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Added
+
+- DDM coordinate format option
+
 ### Changed
 
 - Video: use `media-kit` instead of `ffmpeg-kit` for metadata fetch